% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDefinition.R
\name{addDefinition}
\alias{addDefinition}
\title{addDefinition}
\usage{
addDefinition(itemToDefine, definitionNotes)
}
\arguments{
\item{itemToDefine}{the term you're going to save a definition for}

\item{definitionNotes}{the text you're going to save as the definition. Include
things like a brief decsription of what the problem was, links to the websites
that helped you solve it, the date you're making the defintion and the file names
where your problems occurred.}
}
\value{
successText a message letting you know whether your save was successful or not.
}
\description{
TODO: need to make sure this file is NOT OVERWRITTEN by new updates of the
package.
}
\details{
Our definitions of common R terms are great! But you will run into many more!
This function allows you to add your own terms to the database on your computer.
What does this mean? You can make your own R notes that you can search at any time.
If you're working on a long term project, have you ever had a weird word or concept
you KNOW you're going to forget?

Use addDefinition() and make yourself some notes for later!
}
