% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{defineDB}
\alias{defineDB}
\title{defineDB}
\format{A data frame with 4 rows and 2 variables:
\describe{
  \item{Word}{Terms used commonly in R}
  \item{Definitions}{Definitions of these terms}
  ...
}}
\usage{
defineDB
}
\description{
This is the data offering definitions of common R terms
}
\keyword{datasets}
