% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineMe.R
\name{defineMe}
\alias{defineMe}
\title{defineMe}
\usage{
defineMe(item)
}
\arguments{
\item{item}{the text that the function will search for in the database}
}
\description{
Did you ever wonder what _atomic_ refers to in R?
What's a _closure_ and why is it not _subsettable_?
This function takes the word you're unsure about and searches for a plain
language explanation from our database. There may be links to further
reading. For example, say you are presented with the error: Error in
encourager[4, 5] : object of type 'closure' is not subsettable and you'd
like to know what on earth a closure is. Try defineMe("closure").
}
