% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{encourageDB}
\alias{encourageDB}
\title{encourageDB}
\format{A data frame with 2 rows and 2 variables:
\describe{
  \item{twitterHandle}{the name or twitter handle of the encourager}
  \item{Text}{the encouragement}
  ...
}}
\usage{
encourageDB
}
\description{
This is the data offering encouragement from R users
}
\keyword{datasets}
