% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encourager.R
\name{encourager}
\alias{encourager}
\title{encourager}
\usage{
encourager()
}
\description{
Inspired by the praise function.
A randomly generated assortment of encouragement from the Rstats community
that demonstrates that failure isn't just common when coding - it's essential.
A selection of helpful hints and tips from the Rstats community about their own
pitfalls and difficulties - and how they solved them.

Success isn't easy, we promise we've all been there.
}
\examples{
encourager()

}
