% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{helpDB}
\alias{helpDB}
\title{helpDB}
\format{A data frame with 5 rows and 3 variables:
\describe{
  \item{Error.Message}{The original error message.}
  \item{Common.Form}{An attempt to provide a general form of the message.}
  \item{Help}{A plain text explanation of the message.}

  ...
}}
\usage{
helpDB
}
\description{
This is the data offering plain language explanations of common R errors
}
\keyword{datasets}
