% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpError.R
\name{helpError}
\alias{helpError}
\title{helpError}
\usage{
helpError(item)
}
\arguments{
\item{item}{the text the function will search for in the database}
}
\description{
This function takes a key phrase from the error you don't understand
and looks for a plain language explanation from our database. It may also be
able to provide you with helpful links or tips.
For example, say you are presented with the error:
Error in library(abcd) : there is no package called ‘abcd’
Try using helpError("there is no package called").

Currently search capability is pretty rudimentary. The function works best
searching for short strings of text that don't include the specific variables
that are unique to your workspace.
}
\examples{
helpError("there is no package called")


}
