% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{troubleShootDB}
\alias{troubleShootDB}
\title{troubleShootDB}
\format{A data frame with 2 rows and 2 variables:
\describe{
  \item{twitterHandle}{The name or twitter handle of the R user.}
  \item{Text}{The troubleshooting message.}

  ...
}}
\usage{
troubleShootDB
}
\description{
This is the data offering troubleshooting suggestions from R users
}
\keyword{datasets}
