% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troubleShooter.R
\name{troubleShooter}
\alias{troubleShooter}
\title{troubleShooter}
\usage{
troubleShooter(item)
}
\arguments{
\item{item}{the text the function will search for.}
}
\description{
A function that will generate some troubleshooting tips in response to a limited
number of query items. This will include resources like StackOverflow questions,
websites, links to helpful blog posts or RStudio community posts etc.
}
\examples{
troubleShooter()

}
